smartjournal/requirements.txt
django-extensions
django-crispy-forms==1.14.0
pytesseract
plotly
